Bug fix for logic for codec and fps parameters in Writer.
The logic to handle initiation of Writer did not resolve correctly in scenarios where a multi-image instance of Video was supplied to a movie instance of Writer. This would result in the default codec not being used and None being supplied to cv2.VideoWriter, with a somewhat ambigious exception being raised as a consequence.